Matrices and vectors

+++ lldf-bind/src/types/value/matrix.rs
@@ -5,11 +5,11 @@ use crate::core::ops::*;
 
 #[derive(Clone)]
 #[repr(transparent)]
-pub struct Matrix<const ROWS : u8, const COLUMNS : u8> {
+pub struct Matrix<const ROWS : usize, const COLUMNS : usize> {
     inner : List<Float>
 }
 
-impl<const ROWS : u8, const COLUMNS : u8> Matrix<ROWS, COLUMNS> {
+impl<const ROWS : usize, const COLUMNS : usize> Matrix<ROWS, COLUMNS> {
 
     pub fn zero() -> Self { unsafe {
         #[allow(clashing_extern_declarations)]
@@ -28,7 +28,7 @@ impl<const ROWS : u8, const COLUMNS : u8> Matrix<ROWS, COLUMNS> {
 
 }
 
-impl<const SIZE : u8> Matrix<SIZE, SIZE> {
+impl<const SIZE : usize> Matrix<SIZE, SIZE> {
 
     pub fn identity() -> Self { unsafe {
         #[allow(clashing_extern_declarations)]
@@ -100,7 +100,7 @@ impl Matrix<4, 4> {
 
 }
 
-impl<const ROWS : u8, const COLUMNS : u8> Add<Matrix<ROWS, COLUMNS>> for Matrix<ROWS, COLUMNS> {
+impl<const ROWS : usize, const COLUMNS : usize> Add<Matrix<ROWS, COLUMNS>> for Matrix<ROWS, COLUMNS> {
     type Output = Matrix<ROWS, COLUMNS>;
     fn add(self, rhs : Matrix<ROWS, COLUMNS>) -> Self::Output { unsafe {
         #[allow(clashing_extern_declarations)]
@@ -110,7 +110,7 @@ impl<const ROWS : u8, const COLUMNS : u8> Add<Matrix<ROWS, COLUMNS>> for Matrix<
         let inner = DF_ACTION__SetVariable_CreateList();
         let row = DF_ACTION__Repeat_Multiple(UInt::from(ROWS)); DF_BRACKET__Repeat_Open();
             let column = DF_ACTION__Repeat_Multiple(UInt::from(COLUMNS)); DF_BRACKET__Repeat_Open();
-                let i = column + (row - 1u8) * COLUMNS;
+                let i = column + (row - 1usize) * COLUMNS;
                 let a = DF_ACTION__SetVariable_GetListValue(self.inner, i);
                 let b = DF_ACTION__SetVariable_GetListValue(rhs.inner, i);
                 DF_ACTION__SetVariable_AppendValue(inner.clone(), (a + b).to_opaque());
@@ -120,7 +120,7 @@ impl<const ROWS : u8, const COLUMNS : u8> Add<Matrix<ROWS, COLUMNS>> for Matrix<
     } }
 }
 
-impl<const A_ROWS : u8, const A_COLUMNS_B_ROWS : u8, const B_COLUMNS : u8> Mul<Matrix<A_COLUMNS_B_ROWS, B_COLUMNS>> for Matrix<A_ROWS, A_COLUMNS_B_ROWS> {
+impl<const A_ROWS : usize, const A_COLUMNS_B_ROWS : usize, const B_COLUMNS : usize> Mul<Matrix<A_COLUMNS_B_ROWS, B_COLUMNS>> for Matrix<A_ROWS, A_COLUMNS_B_ROWS> {
     type Output = Matrix<A_ROWS, B_COLUMNS>;
     fn mul(self, rhs : Matrix<A_COLUMNS_B_ROWS, B_COLUMNS>) -> Self::Output { unsafe {
         #[allow(clashing_extern_declarations)]
@@ -129,13 +129,13 @@ impl<const A_ROWS : u8, const A_COLUMNS_B_ROWS : u8, const B_COLUMNS : u8> Mul<M
         }
         let inner = DF_ACTION__SetVariable_CreateList();
         let row = DF_ACTION__Repeat_Multiple(UInt::from(A_ROWS)); DF_BRACKET__Repeat_Open();
-            let a_i0 = (row - 1u8) * A_COLUMNS_B_ROWS;
+            let a_i0 = (row - 1usize) * A_COLUMNS_B_ROWS;
             let column = DF_ACTION__Repeat_Multiple(UInt::from(B_COLUMNS)); DF_BRACKET__Repeat_Open();
                 let cell = DF_TEMPVAR();
                 DF_ACTION__SetVariable_Specialcharequals(cell.clone(), Float::from(0.0));
                 let i = DF_ACTION__Repeat_Multiple(UInt::from(A_COLUMNS_B_ROWS)); DF_BRACKET__Repeat_Open();
                     let a = DF_ACTION__SetVariable_GetListValue(self.inner, a_i0 + i);
-                    let b = DF_ACTION__SetVariable_GetListValue(rhs.inner, column + (i - 1u8) * B_COLUMNS);
+                    let b = DF_ACTION__SetVariable_GetListValue(rhs.inner, column + (i - 1usize) * B_COLUMNS);
                     DF_ACTION__SetVariable_SpecialcharplusSpecialcharequals(cell.clone(), a * b);
                 DF_BRACKET__Repeat_Close();
                 DF_ACTION__SetVariable_AppendValue(inner.clone(), cell);
@@ -145,7 +145,7 @@ impl<const A_ROWS : u8, const A_COLUMNS_B_ROWS : u8, const B_COLUMNS : u8> Mul<M
     } }
 }
 
-impl<const ROWS : u8, const COLUMNS : u8> Matrix<ROWS, COLUMNS> {
+impl<const ROWS : usize, const COLUMNS : usize> Matrix<ROWS, COLUMNS> {
     pub fn transpose(self) -> Matrix<COLUMNS, ROWS> { unsafe {
         #[allow(clashing_extern_declarations)]
         extern "C" {
@@ -154,7 +154,7 @@ impl<const ROWS : u8, const COLUMNS : u8> Matrix<ROWS, COLUMNS> {
         let inner = DF_ACTION__SetVariable_CreateList();
         let row = DF_ACTION__Repeat_Multiple(UInt::from(COLUMNS)); DF_BRACKET__Repeat_Open();
             let column = DF_ACTION__Repeat_Multiple(UInt::from(ROWS)); DF_BRACKET__Repeat_Open();
-                let i = row + (column - 1u8) * COLUMNS;
+                let i = row + (column - 1usize) * COLUMNS;
                 let a = DF_ACTION__SetVariable_GetListValue(self.inner, i);
                 DF_ACTION__SetVariable_AppendValue(inner.clone(), a.to_opaque());
             DF_BRACKET__Repeat_Close();
